backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-Session==0.8.0
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
supabase==2.5.1
requests==2.31.0
jwt==1.3.1